iot/gateway/data_cloud_persister/requirements.txt
paho-mqtt
influxdb-client[ciso]